refactor(invoice): fold CopyForDraftEdit into builder.go
One file: the builder is the mechanism, CopyForDraftEdit is the one named
recipe on top of it — no need for a separate file for a single method.
